+++ srcs/taag/menu_player_editor/menu_choose_sprite.cpp
@@ -103,7 +103,7 @@ void			menu_choose_sprite(t_data data)
 		if (data.data.size() != 0)
 			(*((t_gui *)(data.data[0]))).draw_self();
 		gui.draw_self();
-		if (account->tuto_state < gui_tutorial.size())
+		if ((size_t)(account->tuto_state) < gui_tutorial.size())
 			gui_tutorial[account->tuto_state].draw_self();
 
 
@@ -117,12 +117,12 @@ void			menu_choose_sprite(t_data data)
 				play = false;
 			else if (event.type == SDL_MOUSEBUTTONUP)
 			{
-				if (account->tuto_state < gui_tutorial.size() && gui_tutorial[account->tuto_state].object_list.size() && gui_tutorial[account->tuto_state].click(&event) == true)
+				if ((size_t)(account->tuto_state) < gui_tutorial.size() && gui_tutorial[account->tuto_state].object_list.size() && gui_tutorial[account->tuto_state].click(&event) == true)
 				{
 					increment_tutorial(NULL);
 					gui.click(&event);
 				}
-				else if (account->tuto_state >= gui_tutorial.size())
+				else if ((size_t)(account->tuto_state) >= gui_tutorial.size())
 					gui.click(&event);
 			}
 			else if (event.type == SDL_MOUSEWHEEL && event.wheel.y < 0)
